弹幕实时发送 网关拦截ws请求，解析token(如有)，websocket服务器存储当前会话用户id(如已登录)，弹幕存储和广播

+++ sustech-course-base/sustech-course-interaction/src/main/java/edu/sustech/interaction/entity/Danmu.java
@@ -55,7 +55,7 @@ public class Danmu implements Serializable {
     private String color;
 
     @ApiModelProperty("弹幕所在视频的时间点(分秒，十分之一秒)")
-    private Integer timePoint;
+    private Double timePoint;
 
     @ApiModelProperty("弹幕状态 1默认过审 2被举报审核中 3删除")
     private Byte state;
